+++ Sources/VComponents/Components/Buttons/Rounded Labeled Button/VRoundedLabeledButtonState.swift
@@ -1,5 +1,5 @@
 //
-//  VRoundedLabeledButtonState.swift
+//  VRoundedCaptionedButtonState.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 17.08.22.
@@ -8,9 +8,9 @@
 import Foundation
 import VCore
 
-// MARK: - V Rounded Labeled Button Internal State
+// MARK: - V Rounded Captioned Button Internal State
 /// Enum that represents state, such as `enabled`, `pressed`, or `disabled`.
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-public typealias VRoundedLabeledButtonInternalState = GenericState_EnabledPressedDisabled
+public typealias VRoundedCaptionedButtonInternalState = GenericState_EnabledPressedDisabled